+++ test/sdk_builds_test.dart
@@ -15,37 +15,37 @@ void main() async {
     _dd.close();
   });
 
-  test('get version', () async {
+  test('fetch version', () async {
     const revision = '31822';
     const channel = 'stable';
 
-    final version = await _dd.getVersion(channel, revision);
+    final version = await _dd.fetchVersion(channel, revision);
 
     expect(version.version, Version.parse('1.1.1'));
   });
 
-  test('get version paths', () async {
+  test('fetch version paths', () async {
     const channel = 'stable';
 
-    final paths = await _dd.getVersionPaths(channel).toList();
+    final paths = await _dd.fetchVersionPaths(channel).toList();
 
     expect(paths, contains('channels/stable/release/29803/'));
     expect(paths, contains('channels/stable/release/latest/'));
     expect(paths.length, greaterThanOrEqualTo(28));
   });
 
-  test('getVersions', () async {
-    final versions = await _dd.getVersions('stable');
+  test('fetchVersions', () async {
+    final versions = await _dd.fetchVersions('stable');
 
     expect(versions.length, greaterThanOrEqualTo(27));
   });
 
-  group('getVersion', () {
+  group('fetchVersion', () {
     test('recent git build', () async {
       const channel = 'dev';
       const revision = '1.11.0-dev.5.2';
 
-      final content = await _dd.getVersion(channel, revision) as GitVersionInfo;
+      final content = await _dd.fetchVersion(channel, revision) as GitVersionInfo;
 
       expect(content.ref, '23736d3630da614c655d0569e1ba5af2021b1c61');
       expect(content.version, Version.parse('1.11.0-dev.5.2'));
@@ -56,7 +56,7 @@ void main() async {
       const channel = 'stable';
       const revision = '44672';
 
-      final content = await _dd.getVersion(channel, revision) as SvnVersionInfo;
+      final content = await _dd.fetchVersion(channel, revision) as SvnVersionInfo;
 
       expect(content.revision, 44672);
       expect(content.version, Version.parse('1.9.1'));
@@ -65,7 +65,7 @@ void main() async {
 
     test('old revision with old date format', () async {
       // dev/release/30039 - 0.8.10.8_r30039
-      final content = await _dd.getVersion('dev', '30039') as SvnVersionInfo;
+      final content = await _dd.fetchVersion('dev', '30039') as SvnVersionInfo;
 
       expect(content.revision, 30039);
       expect(content.version, Version.parse('0.8.10-rev.8.30039'));
@@ -73,11 +73,11 @@ void main() async {
     });
   });
 
-  test('getLink', () async {
+  test('downloadLink', () async {
     const channel = 'stable';
     const revision = '44672';
 
-    final content = await _dd.getDownloadLink(
+    final content = await _dd.createDownloadLink(
         channel, revision, 'api-docs/dart-api-docs.zip');
     expect(content.pathSegments, [
       'download',